refactor api to working condition
