+++ api/index.js
@@ -19,7 +19,7 @@ app.use(
   })
 );
 app.use(express.json());
-app.use(express.static(path.join(__dirname, '../../public')));
+app.use(express.static(path.join(__dirname, '../public')));
 app.use('/api', router);
 
 app.listen(port, (err) => {
