+++ package.json
@@ -71,7 +71,7 @@
     "bulma": "^0.9.3",
     "css-loader": "^5.2.7",
     "eslint": "^7.30.0",
-    "eslint-config-airbnb": "^18.2.1",
+    "eslint-config-airbnb-typescript": "^12.3.1",
     "eslint-plugin-import": "^2.23.4",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.24.0",
